initial renaming
renaming a few files to reference this is for the Party Mix Mark II

off the bat some things are working which is nice. it's almost usable. I just don't have the time to teach myself the XML while also trying to do on the rest of my projects. I think I'll start by documenting what works and then going over the Mixxx documentation